+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas3/reporting/model/BedUtilisationReportData.kt
@@ -1,10 +1,10 @@
 package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.reporting.model
 
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.repository.BedUtilisationBedspaceReportData
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.repository.BedUtilisationBookingCancellationReportData
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.repository.BedUtilisationBookingReportData
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.repository.BedUtilisationBookingTurnaroundReportData
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.repository.BedUtilisationVoidBedspaceReportData
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.repository.BedUtilisationBedspaceReportData
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.repository.BedUtilisationBookingCancellationReportData
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.repository.BedUtilisationBookingReportData
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.repository.BedUtilisationBookingTurnaroundReportData
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.repository.BedUtilisationVoidBedspaceReportData
 
 data class BedUtilisationReportData(
   val bedspaceReportData: BedUtilisationBedspaceReportData,
