+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas3/reporting/model/BookingsReportDataAndPersonInfo.kt
@@ -1,6 +1,6 @@
 package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.reporting.model
 
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.repository.BookingsReportData
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.repository.BookingsReportData
 
 data class BookingsReportDataAndPersonInfo(
   val bookingsReportData: BookingsReportData,
