+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas3/repository/BedUsageRepository.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.repository
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.repository
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
